+++ Pipes.h
@@ -8,25 +8,25 @@
 class Pipe : public Object
 {
 private:
-    static int reset_x;
+    static float reset_x;
     sf::Texture t_up, t_down;
     sf::Sprite s_up, s_down;
     std::string file_up, file_down;
 public:
-    Pipe(int _x = 600, std::string _file_up = "images//Pipe_up.png", std::string _file_down = "images//Pipe_down.png");
+    Pipe(float _x = 600, std::string _file_up = "images//Pipe_up.png", std::string _file_down = "images//Pipe_down.png");
     Pipe(const Pipe&);
-    void setX(int val) { x = val; }
-    void setY(int val) { y = val; }
+    void setX(float val) { x = val; }
+    void setY(float val) { y = val; }
     void move() { x--; }
     void reset();
     void draw(sf::RenderWindow&) override;
     void updatePosition();
     bool passed();
     Pipe& operator =(const Pipe&);
-    Pipe operator +(int) const;
-    Pipe operator -(int) const;
-    void operator +=(int);
-    void operator -=(int);
+    Pipe operator +(float) const;
+    Pipe operator -(float) const;
+    void operator +=(float);
+    void operator -=(float);
     void manageExceptions();
 };
